requirements.txt
numpy==1.24.0
pandas==2.0.3
matplotlib==3.8.0
scikit-learn==1.3.0
umap-learn==0.5.3
seaborn==0.13.0